Adapt aiohttp reify implementation to replace internal cached_property
