+++ yarl/_helpers.pyi
@@ -1,0 +1,6 @@
+from typing import Any
+
+class cached_property:
+    def __init__(self, wrapped: Any) -> None: ...
+    def __get__(self, inst: Any, owner: Any) -> Any: ...
+    def __set__(self, inst: Any, value: Any) -> None: ...
